Move dashboard component into own Dashboard folder
