+++ src/components/index.js
@@ -1,3 +1,4 @@
 export * from './Jobs';
 export * from './Toolbar';
 export { default as MultiValuedSelect } from './MultiValuedSelect';
+export { default as Dashboard } from './Dashboard';
